Added working collisions
New collisions with final velocity based on conservation of energy (assumed elastic).
Objects have a tendancy to overlap when the velocity of the larger (mass) object is greater than the smaller object, causes it to mess with the velocity of both objects as well as other objects in contact with them.